cleanup/variable renaming

+++ cypy2/manager.py
@@ -11,7 +11,7 @@ import pandas as pd
 from io import StringIO
 
 from . import file_utils
-from .import file_settings
+from . import file_settings
 
 
 class ActivityManager(object):
@@ -27,29 +27,33 @@ class ActivityManager(object):
         self.strava_metadata = pd.read_csv(os.path.join(strava_dirpath, 'activities.csv'))
 
         if from_cache:
-            self._load_cache()
+            self.parsed_data = self._load_from_cache(self.cache_dirpath)
+        else:
+            # for now, we will let the user call self.parse_all manually
+            pass
 
 
-    def _load_cache(self):
-
-        filepaths = glob.glob(os.path.join(self.cache_dirpath, '*.p'))
+    @staticmethod
+    def _load_from_cache(cache_dirpath):
+        '''
+        Unpickle cached parsed data
+        '''
+        filepaths = glob.glob(os.path.join(cache_dirpath, '*.p'))
         filepath = sorted(filepaths)[-1]
 
         if len(filepaths) > 1:
             print('Warning: multiple pickle files found; loading %s' % filepath.split(os.sep)[-1])
 
         with open(filepath, 'rb') as file:
-            self.parsed_data = pickle.load(file)
+            data = pickle.load(file)
+        return data
 
-
-    def generate_cache_filename(self):
-
-        timestamp = datetime.datetime.now().strftime('%Y-%m-%d')
-        return os.path.join(self.cache_dirpath, '%s_all-parsed-fit-files.p' % timestamp)
 
 
     def parse_all(self):
-
+        '''
+        Parse all of the FIT files appearing in the Strava metadata
+        '''
         data = []
         errors = []
         for ind, row in self.strava_metadata.iterrows():    
@@ -73,7 +77,16 @@ class ActivityManager(object):
 
 
     def to_cache(self):
+        '''
+        Pickle and cache the parsed data
+        '''
+        timestamp = datetime.datetime.now().strftime('%Y-%m-%d')
+        filepath = os.path.join(self.cache_dirpath, '%s_all-parsed-fit-files.p' % timestamp)
 
-        with open(self.generate_cache_filename(), 'wb') as file:
+        if os.path.isfile(filepath):
+            print('Warning: cached data already exists')
+            return
+
+        with open(filepath 'wb') as file:
             pickle.dump(self.parsed_data, file)
